+++ src/services/movies.services.ts
@@ -1,9 +1,10 @@
 import { db } from '../database'
+import { createMovieQuery, getMovieByTitleQuery } from '../database/sqlScipt'
 import { MovieWithCategory } from '../types'
 
 // GET movies services
 export const getMovieByTitle = async (title: string): Promise<MovieWithCategory> => {
-  const queryString: string = 'SELECT movies.*, categories.title as category_title FROM movies JOIN categories ON categories.id = movies.category WHERE movies.title = $1'
+  const queryString = getMovieByTitleQuery
   try {
     const { rows } = await db.query(queryString, [title])
     const result: MovieWithCategory = rows[0]
@@ -17,9 +18,7 @@ export const getMovieByTitle = async (title: string): Promise<MovieWithCategory>
 
 // CREATE movies services
 export const createMovie = async (title: string, description: string, releaseDate: string, category: string): Promise<void> => {
-  let query = 'INSERT INTO movies (title, category, description, release_date)'
-  const subquery = ' SELECT $1, id, $2, $3 FROM categories WHERE title = $4'
-  query += subquery
+  const query = createMovieQuery
   try {
     await db.query(query, [title, description, releaseDate, category])
   } catch (err: any) {