+++ src/services/user.services.ts
@@ -1,11 +1,12 @@
 import { db } from '../database'
+import { createUserQuery, getUserByEmailQuery, getUserByIdQuery, getUserByUsernameQuery, getUsersQuery } from '../database/sqlScipt'
 import { UserFull, UserSecure } from '../types'
 import { cryptPassword } from '../utils/encryptPassword'
 
 // GET SERVICES
 
 export const getUserByUsername = async (username: string): Promise<UserFull > => {
-  const queryString: string = 'SELECT * FROM users WHERE username = $1'
+  const queryString = getUserByUsernameQuery
   try {
     const { rows } = await db.query(queryString, [username])
     const result: UserFull = rows[0]
@@ -17,7 +18,7 @@ export const getUserByUsername = async (username: string): Promise<UserFull > =>
 }
 
 export const getUserByEmail = async (email: string): Promise<UserFull> => {
-  const queryString: string = 'SELECT * FROM users WHERE email = $1'
+  const queryString = getUserByEmailQuery
   try {
     const { rows } = await db.query(queryString, [email])
     const result: UserFull = rows[0]
@@ -29,7 +30,7 @@ export const getUserByEmail = async (email: string): Promise<UserFull> => {
 }
 
 export const getUserById = async (id: number): Promise<UserFull> => {
-  const queryString: string = 'SELECT * FROM users WHERE id = $1'
+  const queryString = getUserByIdQuery
   try {
     const { rows } = await db.query(queryString, [String(id)])
     const result: UserFull = rows[0]
@@ -41,7 +42,7 @@ export const getUserById = async (id: number): Promise<UserFull> => {
 }
 
 export const getUsers = async (): Promise<UserSecure[]> => {
-  const queryString: string = 'SELECT * FROM users'
+  const queryString: string = getUsersQuery
   try {
     const { rows } = await db.query(queryString, [])
     const result: UserSecure[] = rows.map((row: UserFull) => {
@@ -61,7 +62,7 @@ export const getUsers = async (): Promise<UserSecure[]> => {
 // CREATE SERVICES
 
 export const createUser = async (username: string, email: string, password: string): Promise<void> => {
-  const queryString: string = 'INSERT INTO users (username, email, password) VALUES ($1, $2, $3)'
+  const queryString = createUserQuery
   const hash = cryptPassword(password)
   const userParams = [username, email, hash]
   try {