+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# VERSION 0.4.0
+
+## Added
+
+- Get one user with it's watched movies list.
+- todo: Get users list with it's corresponding watched movies.
+- todo: Route that shows movies novelties list.
+
 # VERSION 0.3.0
 
 ## Added